requirements.txt
itchat-uos==1.5.0.dev0
